Attempt to connect a new SSID Bridged AP with current credentials.
This feature adds support to attempt an auto connection to an unsaved
scanned SSID based on the following criteria.

Auto connection attempt shall happen on when.
1. STA is connected currently to a 2G BSSID.
2. Number of BSSIDs scanned for this SSID is less than 6.
3. Connected SSID should not have 5G BSSID in the vicinity.
4. Connected SSID's AKM is WPA/WPA2-PSK.
5. Connected SSID string contain's "2g"/"2G" with some valid
   prefix

The following is the expectation on the unsaved-scanned SSID.

1. New network should be 5G  WPA/WPA2-PSK network
2. New and connected BSSID's shall match the first 16 ascii characters.
   Ex: "c8:b3:73:4c:41:11" AND “c8:b3:73:4c:41:12” matches till
       "c8:b3:73:4c:41:1"
3. New 5G SSID contain's "5g"/"5G" with some valid prefix and further
   should match with that of current connected SSID.

If match found then the new 5G SSID will be created as ephemeral network
with same credentials of connected 2G SSID. Accordingly, this 5G SSID
shall be evaluated for the connection attempt. If the connection
is successful, this 5G SSID is saved, else this 5G SSID / profile is
disabled to avoid retries until Wi-Fi is restarted.

Change-Id: I1a2999a3789d477bf083c0bfeba2b31fe585bc84
CRs-Fixed: 2585822